+++ CHANGELOG.md
@@ -22,8 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Me::username` and `Deref<Target = User>` implementation for `Me` ([#197][pr197])
 - `Me::{mention, tme_url}` ([#197][pr197])
 - `AllowedUpdate::ChatJoinRequest` ([#201][pr201])
+- `ChatId::{is_user, is_group, is_channel_or_supergroup}` functions [#198][pr198]
 
 [pr197]: https://github.com/teloxide/teloxide-core/pull/197
+[pr198]: https://github.com/teloxide/teloxide-core/pull/198
 [pr201]: https://github.com/teloxide/teloxide-core/pull/201
 
 ### Changed
@@ -33,6 +35,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Freeze when getting `RetryAfter(_)` error
   - Retry requests that previously returned `RetryAfter(_)` error
 - `RequestError::RetryAfter` now has a `Duration` field instead of `i32`
+
+### Fixed
+
+- A bug in `Message::url` implementation ([#198][pr198])
 
 ## 0.4.5 - 2022-04-03